requirements.txt
# ============================================================
# Dependencias del Sistema AMZ → ML
# ============================================================

# Core
python-dotenv==1.1.1
requests==2.32.5
requests-aws4auth==1.3.1

# OpenAI (para auto-answer y AI categorías)
openai==2.3.0

# Web UI
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-HTTPAuth==4.8.0
python-socketio==5.10.0

# Process management
psutil==5.9.6

# Database
# (sqlite3 viene con Python)

# Optional but recommended
python-dateutil==2.8.2
pytz==2023.3
